Agregue los metodos showArtistInfoActivity y los metodos que utiliza en MoreDetailsView
